output split
