+++ huginn-net-http/src/lib.rs
@@ -14,3 +14,4 @@ pub mod http_process;
 pub mod display;
 pub mod error;
 pub mod observable;
+pub mod output;
