+++ huginn-net-tcp/src/lib.rs
@@ -3,7 +3,6 @@
 pub use huginn_net_db as db;
 pub use huginn_net_db::tcp;
 
-// TCP processing modules
 pub mod ip_options;
 pub mod mtu;
 pub mod tcp_process;
@@ -14,5 +13,4 @@ pub mod window_size;
 pub mod display;
 pub mod error;
 pub mod observable;
-
-// Placeholder module for future TCP-specific processing built on top of huginn-net-db
+pub mod output;
